refactor: move the fixed-base gadgets into composer::fixed_base
Pure code motion: `component_mul_generator`,
`append_fixed_base_signed_digits`, `assert_canonical_jubjub_scalar` and
the signed-digit width constants move out of `composer.rs` into a
`fixed_base` submodule, together with their soundness tests.
